2/28 Meeting Progress
Added functionality for moving or stationary platforms, which have one-directional collision. Added animated coins which disappear upon pickup. Added functionality for killzones which reload the scene after a short timer, to be used for enemies, spikes, etc. Implemented the killzone in a bottom bound, which also stops the camera from falling endlessly. Added all three enemy animations, although only idle is implemented, and the enemy movement system, which changes the direction of the enemy every time it hits a wall.