3/7 Meeting Progress
Added a killzone to the enemy, as well as behavior to flip the sprite to face the direction its moving. Improved the death effect with slow motion and the player falling through the world. Implemented the jump and run animations, as well as the framework for the death animation. Added player sprite flipping similar to the enemy sprite flip. Added on-screen text hints. Created a game manager which manages the score and will manage the player life status for the death animation.